ai_employee_scripts/pyproject.toml
[project]
name = "ai-employee-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "playwright>=1.58.0",
    "watchdog>=6.0.0",
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "odoorpc>=0.10.1",
    "tweepy>=4.14.0",
    # Platinum tier - OpenAI Agents SDK
    "openai-agents>=0.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "pydantic>=2.0",
    # Deduplication API
    "flask>=3.0.0",
]
